Added 'Overview' link to homepage dropdowns, fixed ToC bug
The table of contents had an issue where setting it to false didn't turn it off. Now fixed.

I also reduced the margins on headings at the top of pages.

On the homepage, sections that have subpages now automatically have an 'Overview' link added to the top of the accordion, so that the title page can be accessed via a link.