Fixing small typos
Important to note - Github scripts linked in text and corresponding figshare entry. If scripts are moved for new Django instance, this will need to be fixed. Links to other parts of the hugo portal instance will also have to be updated in both text and figshare entry for the data