Freya-1887: Remove dependency on publications database (#1278)
* Remove dependency on publications database

Update text on publications dashboard, and remove links to publications database. Need to update Figshare record before this can be merged. Please check for typos and broken links.

* Fixing small typos

Important to note - Github scripts linked in text and corresponding figshare entry. If scripts are moved for new Django instance, this will need to be fixed. Links to other parts of the hugo portal instance will also have to be updated in both text and figshare entry for the data